% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Karatsuba.R
\name{fktimes}
\alias{fktimes}
\title{fktimes - Function for Handling Decimal Multiplication}
\usage{
fktimes(num1, num2)
}
\arguments{
\item{num1}{The first number in string format.}

\item{num2}{The second number in string format.}
}
\value{
The product of the two input numbers with decimal points handled correctly in string format.
}
\description{
This function takes two numbers (which could be integers or decimals in string format) as input and returns their product considering decimal points appropriately.
}
\examples{
\dontrun{
result <- fktimes("-987654.876543", "123.45672345678")
print(result)
}
}
