% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{karatsubaMultiplyRcpp}
\alias{karatsubaMultiplyRcpp}
\title{Karatsuba Multiplication Function in C++}
\usage{
karatsubaMultiplyRcpp(num1, num2)
}
\arguments{
\item{num1}{The first number in string format to be multiplied.}

\item{num2}{The second number in string format to be multiplied.}
}
\value{
The product of the two input numbers in string format.
}
\description{
This function implements the Karatsuba algorithm to multiply two large numbers represented as strings.
}
\examples{
\dontrun{
std::string num1 = "123456789";
std::string num2 = "987654321";
std::string result = karatsubaMultiplyRcpp(num1, num2);
Rcpp::Rcout << result << std::endl;
}
}
